refactor(otlphttp): share response body limit helper
