+++ exporters/otlp/otlplog/otlploghttp/internal/response.go
@@ -1,0 +1,28 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package internal
+
+import (
+	"errors"
+	"fmt"
+	"io"
+)
+
+// CopyResponseBody copies at most maxSize bytes from src to dst. It returns
+// an error if src contains more than maxSize bytes.
+func CopyResponseBody(dst io.Writer, src io.Reader, maxSize int64) error {
+	if _, err := io.Copy(dst, io.LimitReader(src, maxSize)); err != nil {
+		return err
+	}
+
+	var extra [1]byte
+	_, err := io.ReadFull(src, extra[:])
+	if err == nil {
+		return fmt.Errorf("response body too large: exceeded %d bytes", maxSize)
+	}
+	if errors.Is(err, io.EOF) {
+		return nil
+	}
+	return err
+}
