+++ exporters/otlp/otlpmetric/otlpmetrichttp/client_test.go
@@ -498,7 +498,12 @@ func TestResponseBodySizeLimit(t *testing.T) {
 				WithEndpoint(srv.Listener.Addr().String()),
 				WithInsecure(),
 				WithMaxResponseSize(1),
-				WithRetry(RetryConfig{Enabled: false}),
+				WithRetry(RetryConfig{
+					Enabled:         true,
+					InitialInterval: time.Millisecond,
+					MaxInterval:     time.Millisecond,
+					MaxElapsedTime:  time.Second,
+				}),
 			}
 			cfg := oconf.NewHTTPConfig(asHTTPOptions(opts)...)
 			c, err := newClient(cfg)
@@ -532,7 +537,12 @@ func TestResponseBodySizeLimitAfterDecompression(t *testing.T) {
 		WithEndpoint(srv.Listener.Addr().String()),
 		WithInsecure(),
 		WithMaxResponseSize(limit),
-		WithRetry(RetryConfig{Enabled: false}),
+		WithRetry(RetryConfig{
+			Enabled:         true,
+			InitialInterval: time.Millisecond,
+			MaxInterval:     time.Millisecond,
+			MaxElapsedTime:  time.Second,
+		}),
 	}
 	cfg := oconf.NewHTTPConfig(asHTTPOptions(opts)...)
 	c, err := newClient(cfg)
