+++ exporters/otlp/otlpmetric/otlpmetrichttp/internal/response_test.go
@@ -1,0 +1,58 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+package internal
+
+import (
+	"bytes"
+	"errors"
+	"io"
+	"strings"
+	"testing"
+	"testing/iotest"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestCopyResponseBody(t *testing.T) {
+	readErr := errors.New("read error")
+	tests := []struct {
+		name    string
+		src     io.Reader
+		limit   int64
+		want    string
+		wantErr string
+	}{
+		{name: "Empty", src: strings.NewReader(""), limit: 1},
+		{name: "UnderLimit", src: strings.NewReader("x"), limit: 2, want: "x"},
+		{name: "AtLimit", src: strings.NewReader("x"), limit: 1, want: "x"},
+		{
+			name:    "OverLimit",
+			src:     strings.NewReader("xx"),
+			limit:   1,
+			want:    "x",
+			wantErr: "response body too large: exceeded 1 bytes",
+		},
+		{name: "MaxInt64", src: strings.NewReader("ok"), limit: 1<<63 - 1, want: "ok"},
+		{name: "CopyError", src: iotest.ErrReader(readErr), limit: 1, wantErr: readErr.Error()},
+		{
+			name:    "ProbeError",
+			src:     io.MultiReader(strings.NewReader("x"), iotest.ErrReader(readErr)),
+			limit:   1,
+			want:    "x",
+			wantErr: readErr.Error(),
+		},
+	}
+	for _, test := range tests {
+		t.Run(test.name, func(t *testing.T) {
+			var dst bytes.Buffer
+			err := CopyResponseBody(&dst, test.src, test.limit)
+			if test.wantErr == "" {
+				assert.NoError(t, err)
+			} else {
+				assert.EqualError(t, err, test.wantErr)
+			}
+			assert.Equal(t, test.want, dst.String())
+		})
+	}
+}
