+++ exporters/otlp/otlptrace/otlptracehttp/client.go
@@ -210,7 +210,7 @@ func (c *client) UploadTraces(ctx context.Context, protoSpans []*tracepb.Resourc
 			// Success, do not retry.
 			// Read the partial success message, if any.
 			var respData bytes.Buffer
-			if err := copyResponseBody(&respData, resp.Body, c.cfg.MaxResponseSize); err != nil {
+			if err := internal.CopyResponseBody(&respData, resp.Body, c.cfg.MaxResponseSize); err != nil {
 				return err
 			}
 			if respData.Len() == 0 {
@@ -248,7 +248,7 @@ func (c *client) UploadTraces(ctx context.Context, protoSpans []*tracepb.Resourc
 		// message to be returned. It will help in
 		// debugging the actual issue.
 		var respData bytes.Buffer
-		if err := copyResponseBody(&respData, resp.Body, c.cfg.MaxResponseSize); err != nil {
+		if err := internal.CopyResponseBody(&respData, resp.Body, c.cfg.MaxResponseSize); err != nil {
 			return err
 		}
 		respStr := strings.TrimSpace(respData.String())
@@ -385,22 +385,6 @@ type request struct {
 func (r *request) reset(ctx context.Context) {
 	r.Body = r.bodyReader()
 	r.Request = r.WithContext(ctx)
-}
-
-func copyResponseBody(dst io.Writer, src io.Reader, maxSize int64) error {
-	if _, err := io.Copy(dst, io.LimitReader(src, maxSize)); err != nil {
-		return err
-	}
-
-	var extra [1]byte
-	_, err := io.ReadFull(src, extra[:])
-	if err == nil {
-		return fmt.Errorf("response body too large: exceeded %d bytes", maxSize)
-	}
-	if errors.Is(err, io.EOF) {
-		return nil
-	}
-	return err
 }
 
 // retryableError represents a request failure that can be retried.
